Move loading.js one level higher
Also use `template.js` for boundary UI

x-ref: https://vercel.slack.com/archives/C035J346QQL/p1663195222868229?thread_ts=1663171152.925879&cid=C035J346QQL
