+++ app/loading/template.server.tsx
@@ -1,0 +1,6 @@
+import { Boundary } from '@/ui/Boundary.server';
+import React from 'react';
+
+export default function Template({ children }: { children: React.ReactNode }) {
+  return <Boundary>{children}</Boundary>;
+}
